Orange bar header to the Submit page and othe roddities
Not sure if this is right, but I updated the loop-templates/content-page.php to include the orange bar to fit the mock-up.

Also note the unknown ramifications for updating the single-submission.php to include the #primary id to this page to get the width correct. Seemed to work. It needed this #primary id